requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
djangorestframework==3.15.1
drf-spectacular==0.27.2
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.1
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.63
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openapi-codec==1.3.2
orjson==3.10.3
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
simplejson==3.19.2
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
tenacity==8.3.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.3
yarl==1.9.4
